% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMissingFish.R
\name{AddMissingFish}
\alias{AddMissingFish}
\title{AddMissingFish adds in zero observations for missing fish}
\usage{
AddMissingFish(Data)
}
\arguments{
\item{Data}{}
}
\value{
data with missing observations added in
}
\description{
AddMissingFish adds in zero observations for missing fish
}

