% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeAtLength.R
\name{AgeAtLength}
\alias{AgeAtLength}
\title{AgeAtLength 
returns the age at length for a Von Bert relationship}
\usage{
AgeAtLength(Lengths, Fish, Error)
}
\arguments{
\item{Lengths}{a vector of lengths}

\item{Fish}{a list object containing von bert parameters}

\item{Error}{the base standard deviation of the age at length conversion}
}
\value{
a vector of ages at length
}
\description{
AgeAtLength 
returns the age at length for a Von Bert relationship
}

