% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyLifeHistoryError.R
\name{ApplyLifeHistoryError}
\alias{ApplyLifeHistoryError}
\title{ApplyLifeHistoryError
Applies error to life history parameters}
\usage{
ApplyLifeHistoryError()
}
\value{
Fish object with error added in
}
\description{
ApplyLifeHistoryError
Applies error to life history parameters
}

