% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessLBSPREstimates.R
\name{AssessLBSPRResiduals}
\alias{AssessLBSPRResiduals}
\title{Assess the residuals from the LBSPR function}
\usage{
AssessLBSPRResiduals(Estimates, Fish, Year)
}
\arguments{
\item{Estimates}{predicted and observed length data}

\item{Fish}{Fish based life history}

\item{Year}{The current year}
}
\description{
\code{AssessLBSPRResiduals} look at the residuals between observed
and predicted cohorts by LBSPR to look for bias
}

