% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildLBSPRPars.R
\name{BuildLBSPRPars}
\alias{BuildLBSPRPars}
\title{Build par file for LBSPR}
\usage{
BuildLBSPRPars(Fish, Mpow, NGTG, MaxSD, FecB, SL50Min, SL50Max, DeltaMin,
  DeltaMax)
}
\arguments{
\item{Fish}{}

\item{Mpow}{}

\item{NGTG}{}

\item{MaxSD}{}

\item{FecB}{}

\item{SL50Min}{}

\item{SL50Max}{}

\item{DeltaMin}{}

\item{DeltaMax}{}
}
\value{
a pars file for LBSPR ADMB
}
\description{
Build par file for LBSPR
}

