% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCPUE.R
\name{CalculateCPUE}
\alias{CalculateCPUE}
\title{Calculate Catch Per Unit Effort}
\usage{
CalculateCPUE(CPUE, Years, Weights, Form)
}
\arguments{
\item{Form}{}
}
\description{
Calculate Catch Per Unit Effort
}

