% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDensity.R
\name{CalculateDensity}
\alias{CalculateDensity}
\title{Calculate Density}
\usage{
CalculateDensity(Densities, Years, Weights, Form)
}
\arguments{
\item{Form}{}
}
\value{
density weighted by distance to MPA border
}
\description{
Calculate Density
}

