% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchCurve.R
\name{CatchCurve}
\alias{CatchCurve}
\title{Run a Catch Curve}
\usage{
CatchCurve(LengthDat, CatchCurveWeight, WeightedRegression, ReserveYr,
  OutsideBoundYr, ManualM, GroupMPA, Iterations, BootStrap, LifeError,
  HistInterval)
}
\arguments{
\item{HistInterval}{}
}
\value{
An assessment output
}
\description{
Run a Catch Curve
}

