% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDensitySampleSize.R
\name{CheckLengthSampleSize}
\alias{CheckLengthSampleSize}
\title{Check Sample Size}
\usage{
CheckLengthSampleSize(Data)
}
\arguments{
\item{Data}{}
}
\description{
Checks the sample size of the length data to make sure you've got enough
}

