% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPUERatio.R
\name{CPUERatio}
\alias{CPUERatio}
\title{Calculate CPUE Ratios}
\usage{
CPUERatio(CPUEDat, LagLength, Weight, Form, Iterations, BootStrap)
}
\arguments{
\item{BootStrap}{}
}
\value{
an assessment output
}
\description{
\code{CPUERatio} calculates a CPUE ratio in the manner of
}

