% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBSRA.R
\name{DBSRA}
\alias{DBSRA}
\title{Depletion Based Stock Reduction Analysis}
\usage{
DBSRA(CatchDat, DCAC.start.yr, DCAC.end.yr, delta.yr, DBSRA.OFL.yr,
  FMSYtoMRatio, SD.FMSYtoMratio, Delta, SD.Delta, DeltaLowerBound,
  DeltaUpperBound, BMSYtoB0ratio, SD.BMSYtoB0ratio, BMSYtoB0LowerBound,
  BMSYtoB0UpperBound, InterpCatch, Niter)
}
\arguments{
\item{Niter}{}
}
\value{
DBSRA outputs
}
\description{
Depletion Based Stock Reduction Analysis
}

