% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteFiles.R
\name{DeleteFiles}
\alias{DeleteFiles}
\title{Delete ADMB output files}
\usage{
DeleteFiles(ADMBDir)
}
\arguments{
\item{ADMBDir}{An object of class \code{character} that specifies the 
location of the ADMB executable.}
}
\value{
\code{NULL}
}
\description{
Function to delete output files from ADMB (if they exist) and perform cleanup.
}
\examples{
     
\dontrun{

} 

}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

