% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityRatio.R
\name{DensityRatio}
\alias{DensityRatio}
\title{Run Density Ratio}
\usage{
DensityRatio(DenDat, LagLength, Weight, Form, Iterations, BootStrap)
}
\arguments{
\item{BootStrap}{}
}
\value{
DensityRatio assessment
}
\description{
Run Density Ratio
}

