% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBAR.R
\name{LBAR}
\alias{LBAR}
\title{Run LBAR}
\usage{
LBAR(LengthDat, LagLength, Weight, IncludeMPA, ReserveYr, OutsideBoundYr,
  Iterations, BootStrap, LifeError, Lc = NA)
}
\arguments{
\item{Lc}{}
}
\value{
LBAR assessment
}
\description{
Run LBAR
}

