% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBSPR.R
\name{LBSPR}
\alias{LBSPR}
\title{Run LBSPR}
\usage{
LBSPR(LengthDat, EstimateM, Iterations, BootStrap, LifeError, LengthBins,
  ReserveYear, SL50Min, SL50Max, DeltaMin, DeltaMax, IncludeReserve)
}
\arguments{
\item{IncludeReserve}{}
}
\value{
LBSPR Assessment
}
\description{
Run LBSPR
}

