% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaturityAtAge.R
\name{MaturityAtAge}
\alias{MaturityAtAge}
\title{MaturityAtAge}
\usage{
MaturityAtAge(Length, Fish)
}
\arguments{
\item{Fish}{}
}
\value{
Vector of percent sexually mature at age
}
\description{
MaturityAtAge
}

