% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDensityData.R
\name{PlotDensityData}
\alias{PlotDensityData}
\title{Plot Density Data}
\usage{
PlotDensityData(DensityDat, FigureFolder, Fish, Species, Site, Theme)
}
\arguments{
\item{Theme}{}
}
\description{
Plot Density Data
}

