% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLengthData.R
\name{PlotLengthData}
\alias{PlotLengthData}
\title{Plot Length Data}
\usage{
PlotLengthData(LengthDat, FigureFolder, Fish, Species, Site, Theme)
}
\arguments{
\item{Theme}{}
}
\description{
Plot Length Data
}

