% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunLBSPRAssess.R
\name{RunLBSPRAssessFun}
\alias{RunLBSPRAssess}
\alias{RunLBSPRAssessFun}
\title{A wrapper function to execute the ADMB optimization routine}
\usage{
RunLBSPRAssess(AssessPars, LenFreq, LenMids, ADMBDir, ExName = "lbspr",
  showOutput = FALSE, MaxCount = 5, ADMBRead = NULL)
}
\arguments{
\item{AssessPars}{An object of class \code{list} that contains the parameters
required for the LB-SPR assessment. Use \code{LoadAssessPars} function to 
create this object.}

\item{LenFreq}{A \code{numeric} vector that contains the counts of the length
observed data corresponding to the \code{LenMids} length bins Must be the 
same length as \code{LenMids}.}

\item{LenMids}{A \code{numeric} vector that contains the mid-points of the 
length classes corresponding to the \code{LenFreq} data. Must be the same 
length as \code{LenFreq}.}

\item{ADMBDir}{An object of class \code{character} that specifies the 
location of the ADMB executable file.}

\item{ExName}{An object of class \code{character} that specifies the name of 
the ADMB executalbe file. Default is \code{lbspr} and should not include 
extension.}

\item{showOutput}{An object of class \code{logical} that determines if the 
output of the ADMB will be displayed in the R Console. Apparently only 
works on Windows OS.}

\item{MaxCount}{An object of class \code{numeric} that specifies the maximum 
number of times the ADMB code is executed with different starting 
parameters before the determining that the model failed to converge. 
Default is 5.}

\item{ADMBRead}{ADD DETAILS HERE}
}
\value{
A \code{list} that contains: 
\itemize{ 
  \item{\code{ModelFailed}}{ A \code{logical} that determines if the model 
  failed to converge. If #'   \code{TRUE} then other objects in list are 
  likely to be \code{NULL} or  missing completely.} 
  \item{\code{Estimates}}{ A \code{matrix} containing
  the parameter estimates.} 
  \item{\code{Pred}}{ A \code{numeric} vector
  containing the predicted relative size composition.} 
  \item{\code{Obs}}{ A \code{numeric} vector
  containing the observed size composition.}
  \item{\code{Bins}}{ A \code{numeric} vector containing the length class 
  corresonding to observed and predicted length data.} 
  \item{\code{ObjFunVal}}{ A \code{numeric} value specifying the value 
  of the objective function of the ADMB model.} 
  \item{\code{Grad}}{ The final gradient value.} 
  \item{\code{Cormat}}{ A \code{matrix} containing the estimated 
  variance-covariance matrix.} }
}
\description{
This function runs the ADMB assessment code and returns of list of parameter 
estimates, model fits, and some other information.
}
\examples{
     
\dontrun{

}  

}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

