% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity.R
\name{selectivity}
\alias{selectivity}
\title{Calculate Selectivity}
\usage{
selectivity(length, shape, form = "logistic")
}
\arguments{
\item{form}{}
}
\value{
return selectivity at length ogive
}
\description{
Calculate Selectivity
}

