% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimMod_AgeStruc.R
\name{SimMod_AgeStruc}
\alias{SimMod_AgeStruc}
\title{Generate size structure using GTG model and calculate relative YPR and SPR}
\usage{
SimMod_AgeStruc(SimPars, Mpar, ...)
}
\arguments{
\item{SimPars}{An object of class \code{list} that contains all parameters
required to run GTG model.  Full description of model to be added at later
date.}

\item{Mpar}{The natural mortality rate. As this model is age-structured it requires
a value for the mean M of the stock. The other rate parameters (K and F) are calculated
from the ratios provided in \code{SimPars}.}

\item{kslope}{An object of class \code{numeric} which determines the slope of the 
the natural mortality for each GTG to approximate equal fitness across GTG. 
A value of 0 means all GTG have the same natural mortality. The value must be 
quite small to ensure that M is not negative for any groups. See manuscript for 
more details.}
}
\value{
To add details later.
}
\description{
Function to simulate length-structured growth-type-group (GTG) model to 
generate size equilibrium composition of population and catch, as well as SPR
of stock and relative yield. This model is follows the same length-structured GTG 
approach of \code{SimMod_LHR} but uses the rate parameters directly (i.e M, K, F)
rather than the ratios. The results from the two models should be identical if the 
life history and fishing ratios are the same.
}
\author{
Adrian Hordyk
}

