% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimMod_LHR.R
\name{SimMod_LHR}
\alias{SimMod_LHR}
\title{Generate size structure using GTG model and calculate relative YPR and SPR}
\usage{
SimMod_LHR(SimPars, ...)
}
\arguments{
\item{SimPars}{An object of class \code{list} that contains all parameters
required to run GTG model.  Full description of model to be added at later
date.}

\item{kslope}{An object of class \code{numeric} which determines the slope of the 
the natural mortality for each GTG to approximate equal fitness across GTG. 
A value of 0 means all GTG have the same natural mortality. The value must be 
quite small to ensure that M is not negative for any groups. See manuscript for 
more details.}
}
\value{
To add details later.
}
\description{
Function to simulate length-structured growth-type-group (GTG) model to 
generate size equilibrium composition of population and catch, as well as SPR
of stock and relative yield. This model is parameterised with the life-history ratios.
}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

