% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPanel.R
\name{SummaryPanel}
\alias{SummaryPanel}
\title{Plot Summary Panels}
\usage{
SummaryPanel(AssessData, LengthDat, CPUEDat, Species, Site, YearsToSmooth,
  Theme)
}
\arguments{
\item{Theme}{}
}
\description{
Plot Summary Panels
}

