% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDat.R
\name{WriteDat}
\alias{WriteDat}
\title{Write data file (.dat) for ADMB}
\usage{
WriteDat(AssessPars, LengthMids, LenFreq, ADMBDir, ExName = "lbspr")
}
\arguments{
\item{AssessPars}{An object of class \code{list} that contains the neccessary
parameters. Created by \code{LoadAssessPars} function.}

\item{LengthMids}{A vector containing the mid-points of the length bins.
Created by \code{MakeLengthComp}.}

\item{LenFreq}{A vector containing the observed number in each length bins.
Created by \code{MakeLengthComp}.}

\item{ADMBDir}{An object of class \code{character} that specifies the
location of the ADMB executable.}

\item{ExName}{An object of class \code{character} that specifies the name of
the ADMB exectuable (without extension).}
}
\value{
\code{NULL}
}
\description{
A function to write the.dat file required by the ADMB code.
}
\examples{
     
\dontrun{

}  

}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

